+++ lib/extensions/iterable_widget.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/widgets.dart';
+
+extension XIterableWidget on Iterable<Widget> {
+  Widget stack({
+    Key? key,
+    AlignmentGeometry alignment = AlignmentDirectional.topStart,
+    TextDirection textDirection = TextDirection.ltr,
+    StackFit fit = StackFit.loose,
+    Clip clipBehavior = Clip.hardEdge,
+  }) =>
+      Stack(
+        key: key,
+        alignment: alignment,
+        textDirection: textDirection,
+        fit: fit,
+        clipBehavior: clipBehavior,
+        children: _l,
+      );
+
+  Widget row({
+    Key? key,
+    CrossAxisAlignment crossAxisAlignment = CrossAxisAlignment.center,
+    MainAxisAlignment mainAxisAlignment = MainAxisAlignment.start,
+    MainAxisSize mainAxisSize = MainAxisSize.max,
+    VerticalDirection verticalDirection = VerticalDirection.down,
+    TextBaseline textBaseline = TextBaseline.alphabetic,
+    TextDirection textDirection = TextDirection.ltr,
+  }) =>
+      Row(
+        key: key,
+        crossAxisAlignment: crossAxisAlignment,
+        mainAxisAlignment: mainAxisAlignment,
+        mainAxisSize: mainAxisSize,
+        verticalDirection: verticalDirection,
+        textBaseline: textBaseline,
+        textDirection: textDirection,
+        children: _l,
+      );
+
+  Widget column({
+    Key? key,
+    CrossAxisAlignment crossAxisAlignment = CrossAxisAlignment.center,
+    MainAxisAlignment mainAxisAlignment = MainAxisAlignment.start,
+    MainAxisSize mainAxisSize = MainAxisSize.max,
+    VerticalDirection verticalDirection = VerticalDirection.down,
+    TextBaseline textBaseline = TextBaseline.alphabetic,
+    TextDirection textDirection = TextDirection.ltr,
+  }) =>
+      Column(
+        key: key,
+        crossAxisAlignment: crossAxisAlignment,
+        mainAxisAlignment: mainAxisAlignment,
+        mainAxisSize: mainAxisSize,
+        verticalDirection: verticalDirection,
+        textBaseline: textBaseline,
+        textDirection: textDirection,
+        children: _l,
+      );
+
+  List<Widget> get _l => this is List<Widget> ? this as List<Widget> : toList();
+}